+++ src/routes/docs/tomes.ts
@@ -21,6 +21,7 @@ import logos from '$routes/docs/logos/+page.svelte';
 import theming from '$routes/docs/theming/+page.svelte';
 import api from '$routes/docs/api/+page.svelte';
 import intersect from '$routes/docs/intersect/+page.svelte';
+import svelte_preprocess_mdz from '$routes/docs/svelte_preprocess_mdz/+page.svelte';
 import vite_plugin_library_well_known from '$routes/docs/vite_plugin_library_well_known/+page.svelte';
 import mdz from '$routes/docs/mdz/+page.svelte';
 
@@ -35,7 +36,7 @@ export const tomes: Array<Tome> = [
 		name: 'introduction',
 		category: 'guide',
 		Component: introduction,
-		related_tomes: ['api'],
+		related_tomes: ['api', 'svelte_preprocess_mdz'],
 		related_modules: [],
 		related_declarations: [],
 	},
@@ -54,6 +55,22 @@ export const tomes: Array<Tome> = [
 		related_tomes: [],
 		related_modules: [],
 		related_declarations: [],
+	},
+	{
+		name: 'svelte_preprocess_mdz',
+		category: 'helpers',
+		Component: svelte_preprocess_mdz,
+		related_tomes: ['mdz'],
+		related_modules: ['svelte_preprocess_mdz.ts', 'mdz_to_svelte.ts'],
+		related_declarations: ['svelte_preprocess_mdz', 'mdz_to_svelte', 'MdzPrecompiled'],
+	},
+	{
+		name: 'vite_plugin_library_well_known',
+		category: 'helpers',
+		Component: vite_plugin_library_well_known,
+		related_tomes: ['api', 'LibraryDetail', 'LibrarySummary'],
+		related_modules: ['vite_plugin_library_well_known.ts', 'library_gen.ts'],
+		related_declarations: ['vite_plugin_library_well_known', 'library_gen'],
 	},
 	{
 		name: 'csp',
@@ -80,18 +97,10 @@ export const tomes: Array<Tome> = [
 		related_declarations: [],
 	},
 	{
-		name: 'vite_plugin_library_well_known',
-		category: 'helpers',
-		Component: vite_plugin_library_well_known,
-		related_tomes: ['api', 'LibraryDetail', 'LibrarySummary'],
-		related_modules: ['vite_plugin_library_well_known.ts', 'library_gen.ts'],
-		related_declarations: ['vite_plugin_library_well_known', 'library_gen'],
-	},
-	{
 		name: 'mdz',
 		category: 'helpers',
 		Component: mdz,
-		related_tomes: [],
+		related_tomes: ['svelte_preprocess_mdz'],
 		related_modules: ['mdz.ts'],
 		related_declarations: ['Mdz', 'mdz_parse', 'MdzNodeView'],
 	},