+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # @fuzdev/fuz_ui
 
+## 0.183.1
+
+### Patch Changes
+
+- remove @ryanatkn/gro from deps to finish migration to @fuzdev/gro ([599b3b0](https://github.com/fuzdev/fuz_ui/commit/599b3b0))
+
+## 0.183.0
+
+### Minor Changes
+
+- upgrade gro after fuzdev transfer ([1925513](https://github.com/fuzdev/fuz_ui/commit/1925513))
+
+## 0.182.1
+
+### Patch Changes
+
+- improve the `DeclarationDetail` examples rendering ([5193ddd](https://github.com/fuzdev/fuz_ui/commit/5193ddd))
+
+## 0.182.0
+
+### Minor Changes
+
+- add `svelte_preprocess_mdz` ([#111](https://github.com/fuzdev/fuz_ui/pull/111))
+- update peer deps ([#111](https://github.com/fuzdev/fuz_ui/pull/111))
+
 ## 0.181.1
 
 ### Patch Changes
@@ -919,13 +944,13 @@
 
 ### Patch Changes
 
-- make `@ryanatkn/gro` and optional peer dep ([589aef5](https://github.com/ryanatkn/fuz/commit/589aef5))
+- make `@fuzdev/gro` and optional peer dep ([589aef5](https://github.com/ryanatkn/fuz/commit/589aef5))
 
 ## 0.115.0
 
 ### Minor Changes
 
-- upgrade `@ryanatkn/gro@0.133.0` from 0.132.0 ([49b6948](https://github.com/ryanatkn/fuz/commit/49b6948))
+- upgrade `@fuzdev/gro@0.133.0` from 0.132.0 ([49b6948](https://github.com/ryanatkn/fuz/commit/49b6948))
 
 ## 0.114.0
 
@@ -1131,7 +1156,7 @@
 
 ### Minor Changes
 
-- upgrade @ryanatkn/gro@0.119 ([d364646](https://github.com/ryanatkn/fuz/commit/d364646))
+- upgrade @fuzdev/gro@0.119 ([d364646](https://github.com/ryanatkn/fuz/commit/d364646))
 
 ## 0.102.1
 
@@ -1762,7 +1787,7 @@
 ### Minor Changes
 
 - extract contextmenu ([ddb3bf1](https://github.com/ryanatkn/fuz/commit/ddb3bf1))
-- upgrade to @ryanatkn/gro from @feltjs/gro ([e802137](https://github.com/ryanatkn/fuz/commit/e802137))
+- upgrade to @fuzdev/gro from @feltjs/gro ([e802137](https://github.com/ryanatkn/fuz/commit/e802137))
 - upgrade `@ryanatkn/belt@0.15.0` from `0.13.1` ([ddb3bf1](https://github.com/ryanatkn/fuz/commit/ddb3bf1))
 
 ### Patch Changes
@@ -2744,7 +2769,7 @@
 
 ## 0.1.3
 
-- move utils from `@ryanatkn/gro`
+- move utils from `@fuzdev/gro`
   ([#30](https://github.com/ryanatkn/fuz/pull/30))
 
 ## 0.1.2